requirements.txt
altgraph==0.17
certifi==2020.12.5
chardet==4.0.0
Deprecated==1.2.12
elevate==0.1.3
future==0.18.2
gitdb==4.0.5
GitPython==3.1.14
idna==2.10
multi-key-dict==2.0.3
pbr==5.5.1
pefile==2019.4.18
PyGithub==1.54.1
pyinstaller==4.2
pyinstaller-hooks-contrib==2021.1
PyJWT==1.7.1
python-dotenv==0.15.0
python-jenkins==1.7.0
pywin32==300
pywin32-ctypes==0.2.0
requests==2.25.1
six==1.15.0
smmap==3.0.5
urllib3==1.26.4
winshell==0.6
wrapt==1.12.1
